put addreferee and add superviouse to regstration in correct postion
