+++ app/Http/Controllers/RefereeController.php
@@ -133,4 +133,30 @@ class RefereeController extends Controller
         $register=Registration::where('idRegistration',$_SESSION['id_registration'])->first();
         $register->refress()->detach($id);
       }
+      public function addreferee(Request $request)
+      {
+          session_start();
+          for($i=0;$i<sizeof($request->data);$i++)
+          {
+              $refree=Referee::where('arabicName',$request->data[$i]['arabicName'])->first();
+
+
+              $check=DB::table('reports')
+              ->where('idRegistrationF','=', $_SESSION['id_registration'])
+              ->where('idRefereedF','=',$refree->idRefereed)->get();
+              $register=Registration::find($_SESSION['id_registration'])->first();
+              if(empty($request->data[$i]['fileName']))
+              {
+                   $register->refress()->attach($refree->idRefereed);
+
+               }
+               else
+               {
+                  $register->refress()->updateExistingPivot($refree->idRefereed,['URLReport' => $request->data[$i]['fileName']->storePublicly('images')]);
+               }
+
+
+          }
+
+      }
 }
